Sync concurrent access to the requests
If AcceptedRequests() or UnmatchedRequests() is
called during test execution when not all requests
are finished yet it causes a race condition
with write operations that are happening in another
goroutine. To fix this issue a simple r/w mutex
was added to protect acceptedRequests and
unmatchedRequests fields on read/write access.